Remove add-criteria
