--- ui/src/app/cohort-search/misc-ui/index.ts
@@ -1,1 +1,0 @@
-export {AddCriteriaComponent} from './add-criteria.component';
