+++ ui/src/app/cohort-search/search-group-item/search-group-item.component.html
@@ -9,7 +9,7 @@
 
     <button
       class="btn btn-link btn-sm"
-      (click)="alert('edit the item!')">
+      (click)="launchWizard()">
       <clr-icon shape="pencil"></clr-icon>
     </button>
