+++ ui/src/app/cohort-search/search-group-item/search-group-item.component.ts
@@ -28,4 +28,8 @@ export class SearchGroupItemComponent {
   get isRequesting() {
     return this.item.get('isRequesting', false);
   }
+
+  launchWizard() {
+    console.log('Edit the item!');
+  }
 }
